requirements.txt
asgiref==3.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.7
django-appconf==1.0.5
django-ckeditor==6.1.0
django-environ==0.8.1
django-extensions==3.1.3
django-filter==21.1
django-imagekit==4.0.2
django-import-export==2.6.1
django-jazzmin==2.4.8
django-js-asset==1.2.2
django-modeltranslation==0.17.3
django-phonenumber-field==5.2.0
django-taggit==1.5.1
et-xmlfile==1.1.0
graphviz==0.17
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
phonenumberslite==8.12.33
pilkit==2.0
Pillow==8.3.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0